lib/provider: Move the library provider to the library module.
